Add service layer for in-memory repo snapshots
