+++ repo2gpt/__init__.py
@@ -1,0 +1,57 @@
+"""Service layer for repo2GPT."""
+
+from .service import (
+    ALWAYS_INCLUDE_FILENAMES,
+    DEFAULT_CODE_EXTENSIONS,
+    DEFAULT_DIRNAME_DENYLIST,
+    DEFAULT_IGNORE_PATTERNS,
+    DEFAULT_IGNORE_PATTERNS_RAW,
+    DEFAULT_MAX_FILE_BYTES,
+    ConsolidatedChunkResult,
+    ProcessingOptions,
+    RepoProcessingResult,
+    RepoSnapshot,
+    RepoSnapshotChunk,
+    TokenEstimator,
+    collect_repo_snapshot,
+    expand_patterns,
+    join_relative_path,
+    load_gitignore_patterns,
+    load_pattern_file,
+    matches_include,
+    matches_patterns,
+    normalize_relative_path,
+    resolve_skip_paths,
+    should_include_file,
+    should_skip_directory,
+    to_posix_path,
+    write_file_structure_summary,
+)
+
+__all__ = [
+    "ALWAYS_INCLUDE_FILENAMES",
+    "DEFAULT_CODE_EXTENSIONS",
+    "DEFAULT_DIRNAME_DENYLIST",
+    "DEFAULT_IGNORE_PATTERNS",
+    "DEFAULT_IGNORE_PATTERNS_RAW",
+    "DEFAULT_MAX_FILE_BYTES",
+    "ConsolidatedChunkResult",
+    "ProcessingOptions",
+    "RepoProcessingResult",
+    "RepoSnapshot",
+    "RepoSnapshotChunk",
+    "TokenEstimator",
+    "collect_repo_snapshot",
+    "expand_patterns",
+    "join_relative_path",
+    "load_gitignore_patterns",
+    "load_pattern_file",
+    "matches_include",
+    "matches_patterns",
+    "normalize_relative_path",
+    "resolve_skip_paths",
+    "should_include_file",
+    "should_skip_directory",
+    "to_posix_path",
+    "write_file_structure_summary",
+]
