+++ tests/test_filtering.py
@@ -5,7 +5,7 @@ import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from main import (
+from repo2gpt.service import (
     ALWAYS_INCLUDE_FILENAMES,
     DEFAULT_CODE_EXTENSIONS,
     ProcessingOptions,
