+++ tests/test_process_repository_integration.py
@@ -7,7 +7,13 @@ sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
 from main import (  # noqa: E402
     DEFAULT_MAX_FILE_BYTES,
+    build_processing_options,
     process_repository,
+)
+from repo2gpt.service import (  # noqa: E402
+    collect_repo_snapshot,
+    resolve_skip_paths,
+    TokenEstimator,
 )
 
 
@@ -63,3 +69,19 @@ def test_process_repository_generates_expected_outputs(tmp_path: Path):
     assert Path(result.consolidated_chunks[0].path) == consolidated_path
     assert result.consolidated_chunks[0].file_count == 1
     assert not result.token_estimator.enabled
+
+    options = build_processing_options(str(repo_dir), args)
+    skip_relatives = resolve_skip_paths(
+        str(repo_dir), (str(repomap_path), str(consolidated_path))
+    )
+    snapshot = collect_repo_snapshot(
+        str(repo_dir),
+        options,
+        skip_relatives=skip_relatives,
+        token_estimator=TokenEstimator(enabled=False),
+    )
+
+    assert snapshot.repo_map_text.splitlines()[0] == repo_dir.name
+    assert snapshot.repo_map_text == repomap_content
+    assert snapshot.chunks[0].content == consolidated_content
+    assert snapshot.chunks[0].file_count == 1
